requirements.txt
pandas==2.1.4
plotly==5.18.0
pydantic==1.10.13
SQLAlchemy==2.0.24
streamlit_extras==0.3.6
streamlit_pydantic==0.6.0
